pyproject.toml
[tool.poetry]
name = "recplace-test"
version = "0.1.0"
description = ""
authors = ["alexey.zhilyaev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
pydantic = "^2.9.2"
alembic = "^1.13.3"
yarl = "^1.17.1"
pydantic-settings = "^2.6.1"
uvicorn = "^0.32.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
isort = "^5.13.2"
httpx = "^0.27.2"
ruff = "^0.7.2"
sqlalchemy-utils = "^0.41.2"

[tool.isort]
profile = "black"
known_first_party = "src"
known_fastapi = "fastapi"
sections=["FUTURE", "STDLIB", "THIRDPARTY", "FASTAPI", "FIRSTPARTY", "LOCALFOLDER"]
line_length = 80

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

